Added tmp_build_shell script
The tmp_build_shell script "moves" an existing build
directory to a tmpfs filesystem, in order to reduce the wear of SSD
storage because of frequent builds in an edit-save-recompile work mode.

Running this script opens a new shell in a temporary copy of its
argument directory. When the shell exits, the temporary directory is removed
and all build results are lost (the state of the argument directory is
restored to that of before executing the tmp_build_shell script).